requirement.txt
torch
transformers
PyPDF2
gradio
pinecone-client
requests
pdfplumber
PyPDF2
